separation of apis
